By MatKor: kidw0124, Crescend0, yurimmmmm#, contest: Codeforces Round #797 (Div. 3), problem: (E) Price Maximization, Accepted, #, Copy
https: //codeforces.com/contest/1690/submission/178950544

+++ 1102. CF Round #797/Team A - kidw0124, Crescend0, yurimmmmm/E.cpp
@@ -18,8 +18,8 @@ void init();
 int main() {
     init();
 
-    int testCase, presentNum, i, k, left, right, result;
-    vector<int> presents;
+    ll testCase, presentNum, i, k, left, right, result;
+    vector<ll> presents;
 
     cin >> testCase;
     while(testCase--) {
@@ -50,7 +50,6 @@ int main() {
 
         cout << result << '\n';
     }
-
 
     return 0;
 }
